youtube-audio-summarizer/requirements.txt
openai
youtube-transcript-api
requests
python-dotenv

